fefactor :     1. 머지 후 남아있는 경로의 불필요한 폴더/ 파일 정리
